requirements.txt
# Core
discord.py[voice]==2.4.0
python-dotenv==1.0.0
yt-dlp==2024.7.1
beautifulsoup4==4.12.2
requests==2.32.2
selenium==4.34.0
webdriver-manager==4.0.2

# Testing
pytest==8.1.0
pytest-asyncio==0.23.0
pytest-mock==3.12.0

# Optional for hot reload development (prebuilt wheels on 3.11; skip on 3.13+)
watchfiles==0.21.0 ; python_version < "3.13"

# TTS
gTTS==2.5.1

# macOS native TTS (guarded so Windows/Linux skip automatically)
PyObjC==10.3.1 ; sys_platform == "darwin"
pyobjc-framework-Cocoa==10.3.1 ; sys_platform == "darwin"

# Voice compression (discord voice)
PyNaCl==1.5.0

# Spotify downloader (left unpinned; pin if you hit resolve issues)
spotify-dlp
